After Due Date Fix (4 lines deleted)
I accidentally grabbed too many lines in a cut and paste of some queries (at about line 334). And I grabbed an IF statement that crashed the python due to a lack of indentation. The IF statement lines were removed along with redundant variable assignments that had also been grabbed in the cut and paste.